ci: make Wayland terminal scroll probe use xterm buffer
