+++ .github/workflows/linux-wayland-gpu-sandbox.yml
@@ -4,6 +4,7 @@ on:
   pull_request:
     paths:
       - .github/workflows/linux-wayland-gpu-sandbox.yml
+      - config/scripts/linux-wayland-renderer-diagnostics.mjs
       - config/scripts/verify-linux-wayland-gpu-sandbox.mjs
       - src/main/startup/configure-process.ts
       - src/main/startup/configure-process.test.ts
